Implement auth service
