+++ report_client/src/services/index.js
@@ -1,1 +1,1 @@
-export * from './login';
+export * from './AuthService';
